Finish section23

+++ admin/functions.php
@@ -1,12 +1,140 @@
 <?php
 
-function confirm_query($the_query) {
+function confirm_query( $query ) {
   global $connection;
-  if ( !$the_query) {
-    die("Query Failed" . mysqli_error($connection));
+  if ( !$query) {
+    die("Query Failed" . mysqli_error($connection) );
   }
-  
 }
+function query_result( $query ) {
+  global $connection;
+  $query_result = mysqli_query ( $connection, $query );
+  confirm_query( $query_result );
+  return $query_result;
+}
+
+function escape_string( $string ) {
+  global $connection;
+  return mysqli_real_escape_string ( $connection, $string );
+}
+
+function update_user_name_email($user_id, $user_info, $new_user_info, $column, $messages) {
+  global $connection;
+  global $messages;
+  if ( $user_info !== $new_user_info ) {
+    $check_user_info_query = "SELECT $column FROM users WHERE $column = '$new_user_info'";
+    $check_user_info_query_result = query_result( $check_user_info_query );
+    if ( mysqli_num_rows( $check_user_info_query_result ) == 0 ) {
+      $update_user_info_query = "UPDATE users SET $column = '$new_user_info' WHERE user_id = $user_id";
+      $update_user_info_query_result = mysqli_query( $connection, $update_user_info_query );
+      confirm_query( $update_user_info_query_result );
+      $_SESSION[ $column ] = $new_user_info;
+    } else {
+      if ( $column === 'user_name' ) {
+        array_push( $messages, "This username already exists");
+      } else {
+        array_push( $messages, "This email already exists");
+      }
+    }
+  }
+}
+
+function update_user_info($user_id, $user_info, $new_user_info, $column) {
+  global $connection;
+  if ( $user_info !== $new_user_info) {
+    $update_user_info_query = "UPDATE users SET $column = '$new_user_info' WHERE user_id = $user_id";
+    $update_user_info_query_result = mysqli_query( $connection, $update_user_info_query );
+    confirm_query( $update_user_info_query_result );
+    $_SESSION[ $column ] = $new_user_info;
+  }
+}
+
+function update_post( $post_id, $post_category_id, $post_title, $post_image, $post_content, $post_status ) {
+  global $connection;
+  $query = "UPDATE posts SET ";
+  $query .= "post_category_id = '$post_category_id', ";
+  $query .= "post_title = '{$post_title}', ";
+  $query .= "post_date = now(), ";
+  $query .= "post_image = '{$post_image}', ";
+  $query .= "post_content = '{$post_content}', ";
+  $query .= "post_status = '{$post_status}' ";
+  $query .= "WHERE post_id = {$post_id} ";    
+
+  $query_result = query_result( $query );
+  // header("Location: posts.php");
+  echo "<h3 class='bg-success'>Post has been updated.</h3>
+  <h4>Click <a href='../post.php?post_id=$post_id'>here</a> to view current post or <a href='./posts.php'>here</a> to view all posts.</h4>";
+
+}
+
+function get_post_info( $post_id ) {
+  $query = "SELECT * FROM posts WHERE post_id = $post_id";
+  $query_result = query_result( $query );
+  $row = mysqli_fetch_assoc( $query_result );
+  $post_array = array(
+    "post_category_id" => $row['post_category_id'],
+    "post_title" => $row['post_title'],
+    "post_author_id" => $row['post_author_id'],
+    "post_date" => $row['post_date'],
+    "post_image" => $row['post_image'],
+    "post_content" => $row['post_content'],
+    "post_status" => $row['post_status'],
+  );
+
+  return $post_array;
+
+}
+
+function delete_query( $table, $column, $value ) {
+  $query = "DELETE FROM $table WHERE $column = $value";
+  $query_result = query_result( $query );
+}
+
+function delete_category() {
+  global $connection;
+  if( isset( $_GET['delete'] )) {
+    $cat_id = $_GET['delete'];
+    delete_query( 'categories', 'cat_id', $cat_id );
+    $select_all_posts_query = "SELECT * FROM posts WHERE post_category_id = $cat_id";
+    $select_all_posts_query_result = query_result( $select_all_posts_query );
+    while ( $row = mysqli_fetch_assoc( $select_all_posts_query_result ) ) {
+      $update_category_query = "UPDATE posts SET post_category_id = 1";
+      $update_category_query_result = query_result( $update_category_query );
+    }
+    header("Location: categories.php");
+  }
+}
+
+
+function delete_post() {
+  global $connection;
+  if ( isset( $_GET['delete'] ) ) {
+    $post_id = $_GET['delete'];
+    delete_query( 'posts', 'post_id', $post_id );
+    delete_post_comments( $post_id );
+    header("Location: posts.php");
+  }
+}
+
+function delete_post_comments( $post_id ) {
+  global $connection;
+  $select_all_comments_querry = "SELECT comment_id FROM comments WHERE comment_post_id = $post_id";
+  $select_all_comments_querry_result = query_result( $select_all_comments_querry );
+  while ( $row = mysqli_fetch_assoc( $select_all_comments_querry_result ) ) {
+    delete_query( 'comments ', 'comment_id', $row['comment_id'] );
+  }
+
+}
+
+function delete_comment() {
+  global $connection;
+  if ( isset( $_GET['delete'] )) {
+    $comment_id = $_GET['delete'];
+    delete_query( 'comments ', 'comment_id', $comment_id );
+    header("Location: comments.php"); 
+  }
+}
+
 
 
 
@@ -31,7 +159,6 @@ function insert_categories() {
   }
 }
 
-
 function find_all_categories() {
   global $connection;
   $query = 'SELECT * FROM categories';
@@ -44,99 +171,49 @@ function find_all_categories() {
       echo "<tr>";
       echo "<td>{$cat_id}</td>";
       echo "<td>{$cat_title}</td>";
-      echo "<td><a href='categories.php?edit={$cat_id}'>EDIT</a></td>"; 
-      echo "<td><a href='categories.php?delete={$cat_id}' OnClick=\"return confirm( 'Are you sure you want to delete this category?' );\">DELETE</a></td>";
+      if ( $cat_id > 1 ) {
+        echo "<td><a href='categories.php?edit={$cat_id}'>EDIT</a></td>"; 
+        echo "<td><a href='categories.php?delete={$cat_id}' OnClick=\"return confirm( 'Are you sure you want to delete this category?' );\">DELETE</a></td>";
+      }
       echo "</tr>";
   }
-
-
 }
 
-function delete_category() {
-  global $connection;
-
-  if( isset( $_GET['delete'] )) {
-    $the_cat_id = $_GET['delete'];
-    $query = "DELETE FROM categories WHERE cat_id = {$the_cat_id}";
-    $delete_query = mysqli_query( $connection, $query );
-    header("Location: categories.php");
+function display_categories( $selected_cat_id ) {
+  $select_all_categories_query = "SELECT * FROM categories";
+  $select_all_categories_query_result = query_result( $select_all_categories_query );
+  while ( $row = mysqli_fetch_assoc( $select_all_categories_query_result ) ) {
+    $cat_id = $row['cat_id'];
+    $cat_title = $row['cat_title'];
+    if ( $cat_id == $selected_cat_id ) {
+      echo "<option selected='selected' value='{$cat_id}'>{$cat_title}</option>";
+    } else {
+      echo "<option value='{$cat_id}'>{$cat_title}</option>";
+    }
   }
-
-}
-
-function delete_post() {
-  global $connection;
-  if ( isset( $_GET['delete'] ) ) {
-    $delete_post_id = $_GET['delete'];
-
-    $query = "DELETE FROM posts WHERE post_id = $delete_post_id";
-    $delete_query = mysqli_query($connection, $query);
-    confirm_query($delete_query);
-    header("Location: posts.php");
-  }
-}
-
-function delete_comment() {
-  global $connection;
-  if ( isset( $_GET['delete'] )) {
-    $delete_Comment_id = $_GET['delete'];
-
-    $delete_comment_query = "DELETE FROM comments WHERE comment_id = $delete_Comment_id";
-    $delete_comment_query_result = mysqli_query( $connection, $delete_comment_query );
-    confirm_query( $delete_comment_query_result );
-
-    header("Location: comments.php");
-
-  }
-}
-
-
-function update_post( $post_id, $post_category_id, $post_title, $post_author, $post_image, $post_content, $post_tags, $post_status ) {
-  global $connection;
-  $query = "UPDATE posts SET ";
-  $query .= "post_category_id = '$post_category_id', ";
-  $query .= "post_title = '{$post_title}', ";
-  $query .= "post_author = '{$post_author}', ";
-  $query .= "post_date = now(), ";
-  $query .= "post_image = '{$post_image}', ";
-  $query .= "post_content = '{$post_content}', ";
-  $query .= "post_tags = '{$post_tags}', ";
-  $query .= "post_status = '{$post_status}' ";
-  $query .= "WHERE post_id = {$post_id} ";    
-
-  $query_update_post = mysqli_query( $connection, $query );
-  confirm_query( $query_update_post );
-  // header("Location: posts.php");
-  echo "<h3 class='bg-success'>Post has been updated.</h3>
-  <h4>Click <a href='../post.php?post_id=$post_id'>here</a> to view current post or <a href='./posts.php'>here</a> to view all posts.</h4>";
-
 }
 
 function update_post_category() {
   global $connection;
   if ( isset( $_POST['update_post_category'] ) && isset( $_GET['post_id'] ) ) {
-    $post_category = $_POST['update_post_category'];
+    $post_status = $_POST['update_post_category'];
     $post_id = $_GET['post_id'];
-    
-    $update_post_category_query = "UPDATE posts SET post_category_id = '$post_category' WHERE  post_id = $post_id";
-    $update_post_category_query_result = mysqli_query( $connection, $update_post_category_query );
-    confirm_query( $update_post_category_query_result );
-    header("Location: posts.php");
 
+    // $update_post_status_query = "UPDATE posts SET post_status = '$post_status' WHERE  post_id = $post_id";
+    // $update_post_status_query_result = query_result( $update_post_status_query );
+    // header("Location: posts.php");
   }
+
 }
 
 function update_post_status() {
   global $connection;
-  if ( isset( $_POST['post_status_update'] ) && isset( $_GET['post_id'] ) ) {
-    $post_status = $_POST['post_status_update'];
-    $post_id = $_GET['post_id'];
-    
+  if ( isset( $_POST['update_post_status'] ) && isset( $_GET['post_id'] ) ) {
+    echo $post_status = $_POST['update_post_status'];
+    echo $post_id = $_GET['post_id'];
     $update_post_status_query = "UPDATE posts SET post_status = '$post_status' WHERE  post_id = $post_id";
-    $update_post_status_query_result = mysqli_query( $connection, $update_post_status_query );
-    confirm_query( $update_post_status_query_result );
+    $update_post_status_query_result = query_result( $update_post_status_query );
     header("Location: posts.php");
-
   }
 
 }
@@ -178,7 +255,6 @@ function delete_user() {
   global $connection;
   if ( isset( $_GET['delete'] ) ) {
     $delete_user_id = $_GET['delete'];
-
     $delete_user_query = "DELETE FROM users WHERE user_id = $delete_user_id";
     $delete_user_query_result = mysqli_query($connection, $delete_user_query);
     confirm_query($delete_user_query_result);
@@ -204,62 +280,21 @@ function update_user_role() {
 
 }
 
-function update_user( $user_id, $user_name, $user_password, $user_firstname, $user_lastname, $user_email, $user_image, $user_role ) {
-  global $connection;
-
-  $randSalt = get_randSalt();
-
-  $user_password = crypt( $user_password, $randSalt );
-
-
-  $update_user_data_query = "UPDATE users SET ";
-  $update_user_data_query .= "user_name = '$user_name', ";
-  $update_user_data_query .= "user_password = '$user_password', ";
-  $update_user_data_query .= "user_firstname = '$user_firstname', ";
-  $update_user_data_query .= "user_lastname = '$user_lastname', ";
-  $update_user_data_query .= "user_email = '$user_email', ";
-  $update_user_data_query .= "user_image = '$user_image', ";
-  $update_user_data_query .= "user_role = '$user_role' ";
-  $update_user_data_query .= "WHERE user_id = $user_id";
-  
-  $update_user_data_query_result = mysqli_query( $connection, $update_user_data_query );
-  confirm_query( $update_user_data_query_result );
-
-  if ( isset( $_SESSION['user_name'] ) && $_SESSION['user_name'] === $user_name ) {
-    $_SESSION['user_name'] = $user_name;
-    $_SESSION['user_firstname'] = $user_firstname;
-    $_SESSION['user_lastname'] = $user_lastname;
-    $_SESSION['user_role'] = $user_role;
-  }
-
-  header("Location: users.php");
-  
-}
-
-function get_randSalt() {
-  global $connection;
-  $get_randSalt_query = "SELECT randSalt FROM users";
-  $get_randSalt_query_result = mysqli_query( $connection, $get_randSalt_query );
-  confirm_query( $get_randSalt_query_result );
-  $row = mysqli_fetch_assoc( $get_randSalt_query_result );
-  return $row['randSalt'];
-}
-
-function set_new_session( $user_id, $user_name, $user_firstname, $user_lastname, $user_role) {
+function set_new_session( $user_id, $user_name, $user_firstname, $user_lastname, $user_email, $user_image, $user_role) {
   global $connection;
   $_SESSION['user_id'] = $user_id;
   $_SESSION['user_name'] = $user_name;
   $_SESSION['user_firstname'] = $user_firstname;
   $_SESSION['user_lastname'] = $user_lastname;
+  $_SESSION['user_email'] = $user_email;
+  $_SESSION['user_image'] = $user_image;
   $_SESSION['user_role'] = $user_role;
-
 }
 
 function get_author_name( $post_author_id ) {
   global $connection;
   $get_author_name_query = "SELECT user_name FROM users WHERE user_id = $post_author_id";
-  $get_author_name_query_result = mysqli_query( $connection, $get_author_name_query );
-  confirm_query( $get_author_name_query_result );
+  $get_author_name_query_result = query_result( $get_author_name_query );
   $row = mysqli_fetch_array( $get_author_name_query_result );
   return $row['user_name'];
 }
@@ -267,17 +302,16 @@ function get_author_name( $post_author_id ) {
 function get_post_category( $post_id ) {
   global $connection;
   $get_post_category_query = "SELECT cat_title FROM categories WHERE cat_id = $post_id";
-  $get_post_category_query_result = mysqli_query( $connection, $get_post_category_query );
-  confirm_query( $get_post_category_query_result );
+  $get_post_category_query_result = query_result( $get_post_category_query );
   $row = mysqli_fetch_array( $get_post_category_query_result );
   return $row['cat_title'];
 }
+
 function get_post_comments_count ( $post_id ) {
   global $connection;
   $approved_comments = 0;
   $get_post_comments_count_query = "SELECT comment_status FROM comments WHERE comment_post_id = $post_id";
-  $get_post_comments_count_query_result = mysqli_query( $connection, $get_post_comments_count_query );
-  confirm_query( $get_post_comments_count_query_result );
+  $get_post_comments_count_query_result = query_result( $get_post_comments_count_query );
   while ( $row = mysqli_fetch_array( $get_post_comments_count_query_result ) ) {
     if ( $row['comment_status'] === 'approved' ) {
       $approved_comments++;
@@ -291,6 +325,8 @@ function get_post_comments_count ( $post_id ) {
 function create_single_post( $post_id, $post_category_id, $post_category_name, $post_title, $post_author_id, $post_author_name, 
                              $post_date, $post_image, $post_content ) {
   ?>
+
+  <!-- HTML -->
   <div class="single-post">
       
       <!-- First Blog Post -->
@@ -339,8 +375,6 @@ function create_single_post( $post_id, $post_category_id, $post_category_name, $
         }
 
       ?>
-
-
     </div>
 
   <?php
@@ -348,9 +382,7 @@ function create_single_post( $post_id, $post_category_id, $post_category_name, $
 
 function display_posts( $query ) {
   global $connection;
-
-  $query_result = mysqli_query( $connection, $query );
-  confirm_query( $query_result );
+  $query_result = query_result( $query );
 
   while ( $row = mysqli_fetch_assoc( $query_result )) {
     $post_id = $row['post_id'];
@@ -374,7 +406,7 @@ function display_posts( $query ) {
   }
 
   if ( mysqli_num_rows( $query_result ) == 0) {
-      echo "<h1>No Posts Yet</h1>";
+    echo "<h1>No Posts Yet</h1>";
   }
 }
 
@@ -404,8 +436,7 @@ function display_comments( $post_id ) {
 
   <?php
   $select_all_comments_querry = "SELECT * from comments WHERE comment_post_id = $post_id AND comment_status = 'approved' ORDER BY comment_id DESC";
-  $select_all_comments_querry_result = mysqli_query( $connection, $select_all_comments_querry );
-  confirm_query( $select_all_comments_querry_result );
+  $select_all_comments_querry_result = query_result( $select_all_comments_querry );
   
   while ( $row = mysqli_fetch_assoc( $select_all_comments_querry_result ) ) {
     $comment_author = $row['comment_author'];
@@ -429,8 +460,6 @@ function display_comments( $post_id ) {
   <?php 
   } 
 
-  
-
 }
 
 function leave_comment() {
@@ -448,21 +477,10 @@ function leave_comment() {
       $insert_comment_query .= "VALUES({$post_id}, '{$comment_author}', '{$comment_content}', ";
       $insert_comment_query .= "'{$comment_email}', 'unapprove', now())";
 
-      $insert_comment_query_result = mysqli_query( $connection, $insert_comment_query );
-
-      confirm_query( $insert_comment_query_result );
+      $insert_comment_query_result = query_result( $insert_comment_query );
       echo "<script>alert('Your message has been registered')</script>";
-
-      
-
   } else {
     echo "<script>alert('The fields cannot be empty')</script>";
-    
   }
- 
 }
-
-
-
-
 ?>
